fix: Use more intuitive error path format (#6)
* feat: StructError to include the full path

When investigating the reason that a large XML failed the validation,
looking at only the NodeName might not be enough since there can be too
many elements with the same name, for example `address`.

At the same time, while the error message contains the actual value, so
it could be used to identify where the issue lies, it may not be
possible to log since that might contain sensitive data.

For that reason a new field is added in StructError that contains the
full path to the node that fails the validation.

Use format root/element1/element2/leaf for path format.

Added a small test utility that has essentially the same implementation
as TestPathInError_RootOverflow, but with out the test setup complexity.
Ran said utility through valgrind with:

valgrind --track-origins=yes --leak-check=full --show-leak-kinds=definite ./util

This reports no errors outside of the "usual" Go runtime stuff. Tested
adding errors (too small allocation, missing free, etc.) to verify that
the valgrind test would report problems.

* fix: allocate a multiple of 32-bytes for path

Looking at valgrind output, there was an invalid read of size 32, which
appears to be in Go internals, probably when it's making a copy of the
string in `C.GoString`. Looks like it's using XMM registers to copy 32
bytes at a time, but doesn't care if it reads beyond the end of allocated
memory.

These changes ensure that copying the string 32 bytes at a time won't
lead to any invalid reads.

Co-authored-by: Georgios Koutsoumpakis <georgios.koutsoumpakis@form3.tech>